imported action to clear movie state in redux on mount...
this actions is sent on mount to clear out any movie data currently in state. This is done so there isn't duplicate results displayed when navigating back to a route. The first page of movies is fetched on mount so if it wasn't cleared out you would have those results getting added to end of existing state.